refactor: move formatting utility into the domain
